feat: Add Construct module for DevSuite Utils
This commit adds the `Construct` module to the DevSuite Utils library. The `Construct` module provides utility methods for constructing various components used in the DevSuite library. It includes the `Config` module for configuration-related functionality.

This feature enhances the functionality of the DevSuite Utils library by providing a centralized module for constructing components, improving code organization and reusability.

Note: This commit does not include any tests for the new functionality.
