+++ lib/dev_suite/directory_tree/config/config.rb
@@ -1,0 +1,12 @@
+# frozen_string_literal: true
+
+module DevSuite
+  module DirectoryTree
+    module Config
+      include Utils::Construct::Config
+
+      require_relative "configuration"
+      require_relative "settings"
+    end
+  end
+end
