+++ lib/dev_suite/directory_tree/renderer/base.rb
@@ -4,12 +4,6 @@ module DevSuite
   module DirectoryTree
     module Renderer
       class Base
-        attr_reader :settings
-
-        def initialize(settings: Settings.new)
-          @settings = settings
-        end
-
         def render
           raise NotImplementedError
         end
