--- lib/dev_suite/directory_tree/settings.rb
@@ -1,18 +1,0 @@
-# frozen_string_literal: true
-
-module DevSuite
-  module DirectoryTree
-    class Settings
-      include Utils::ConfigTools::Settings
-
-      def default_settings
-        {
-          skip_hidden: false,
-          skip_types: [],
-          max_depth: nil,
-          max_size: 100 * 1024 * 1024, # 100 MB
-        }
-      end
-    end
-  end
-end
