--- lib/dev_suite/utils/config_tools.rb
@@ -1,10 +1,0 @@
-# frozen_string_literal: true
-
-module DevSuite
-  module Utils
-    module ConfigTools
-      require_relative "config_tools/configuration"
-      require_relative "config_tools/settings"
-    end
-  end
-end
