+++ lib/dev_suite/utils/construct/config/base_configuration.rb
@@ -2,69 +2,41 @@
 
 module DevSuite
   module Utils
-    module ConfigTools
-      module Configuration
-        # Module for global configuration
-        class << self
-          def included(base)
-            base.extend(ClassMethods)
-            base.include(InstanceMethods)
-          end
-        end
+    module Construct
+      module Config
+        class BaseConfiguration
+          class << self
+            # Defines a configuration attribute with an optional default value
+            # This method automatically creates getter and setter methods for the attribute
+            #
+            # @param attr_name [Symbol] The name of the attribute
+            # @param default_value [Object] The default value for the attribute
+            def config_attr(attr_name, default_value: nil)
+              config_attrs << [attr_name, default_value]
 
-        module ClassMethods
-          # Provides global access to a single instance of the configuration class
-          # @return [Object] The singleton configuration instance
-          def configuration
-            @configuration ||= new
-          end
+              # Getter method for the attribute
+              define_method(attr_name) do
+                instance_variable_get("@#{attr_name}")
+              end
 
-          # Allows block-based configuration
-          # Example usage:
-          #   ConfigClass.configure do |config|
-          #     config.attr_name = value
-          #   end
-          def configure
-            yield(configuration)
-          rescue StandardError => e
-            Utils::ErrorHandler.handle_error(e)
-          end
-
-          # Defines a configuration attribute with an optional default value
-          # This method automatically creates getter and setter methods for the attribute
-          #
-          # @param attr_name [Symbol] The name of the attribute
-          # @param default_value [Object] The default value for the attribute
-          def config_attr(attr_name, default_value: nil)
-            config_attrs << [attr_name, default_value]
-
-            # Getter method for the attribute
-            define_method(attr_name) do
-              instance_variable_get("@#{attr_name}")
+              # Setter method for the attribute with validation and resolution
+              define_method("#{attr_name}=") do |value|
+                validate_attr!(attr_name, value)
+                instance_variable_set("@#{attr_name}", resolve_attr(attr_name, value))
+              end
             end
 
-            # Setter method for the attribute with validation and resolution
-            define_method("#{attr_name}=") do |value|
-              validate_attr!(attr_name, value)
-              instance_variable_set("@#{attr_name}", resolve_attr(attr_name, value))
+            # Stores configuration attributes and their default values
+            # @return [Hash] A hash of attribute names and their default values
+            def config_attrs
+              @config_attrs ||= []
+            end
+
+            def config_attr_present?(attr)
+              config_attrs.any? { |key, _default| key == attr }
             end
           end
 
-          # Stores configuration attributes and their default values
-          # @return [Hash] A hash of attribute names and their default values
-          def config_attrs
-            @config_attrs ||= []
-          end
-
-          # Resets the configuration by reinitializing it with default values
-          def reset!
-            @configuration = nil
-            configuration # Reinitialize the configuration instance
-          end
-        end
-
-        # Module for instance-level methods
-        module InstanceMethods
           private
 
           # Initializes configuration attributes with provided or default values
@@ -91,7 +63,7 @@ module DevSuite
           # @param attr [Symbol] The name of the attribute
           # @param value [Object] The value to validate
           def validate_attr!(attr, value)
-            raise ArgumentError, "Unknown attribute: #{attr}" unless self.class.config_attrs.key?(attr)
+            raise ArgumentError, "Unknown attribute: #{attr}" unless self.class.config_attr_present?(attr)
             raise ArgumentError, "Invalid #{attr} value: #{value}" if value.nil?
 
             # Override in including class to add custom validation logic
@@ -136,7 +108,7 @@ module DevSuite
           # @param value [Object] The value to resolve
           # @return [Object] The resolved value
           def resolve_attr(attr, value)
-            raise ArgumentError, "Unknown attribute: #{attr}" unless self.class.config_attrs.key?(attr)
+            raise ArgumentError, "Unknown attribute: #{attr}" unless self.class.config_attr_present?(attr)
 
             # Override in including class to resolve symbols to objects
             value
