+++ lib/dev_suite/utils/construct/construct.rb
@@ -1,0 +1,9 @@
+# frozen_string_literal: true
+
+module DevSuite
+  module Utils
+    module Construct
+      require_relative "config"
+    end
+  end
+end
