+++ lib/dev_suite/utils/table/renderer/simple.rb
@@ -19,6 +19,10 @@ module DevSuite
 
           private
 
+          def settings
+            Config.configuration.settings
+          end
+
           #
           # Calculates the widths of the columns
           #
