+++ lib/dev_suite/utils/utils.rb
@@ -5,7 +5,7 @@ module DevSuite
     require_relative "emoji"
     require_relative "logger"
     require_relative "error_handler"
-    require_relative "config_tools"
+    require_relative "construct"
     require_relative "color"
     require_relative "table"
     require_relative "file_loader"
